requirements.txt
# Core environment
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
numpy==1.26.4

# 🧠 AI + LLM integrations
openai==1.47.1
anthropic==0.30.0
google-generativeai==0.8.3

# --- Database + API ---
supabase>=2.4.3
postgrest>=0.13.1
gotrue>=2.12.0
storage3>=0.7.4

# --- HTTP + async requests ---
httpx>=0.25,<1


# 🧰 Automation + Scheduling
apscheduler==3.10.4
tenacity==8.3.0

# 📊 Visualization / Dashboard (for Retool compatibility)
plotly==5.24.0

# 🔐 Security + logging
cryptography==43.0.1
python-json-logger==2.0.7
loguru==0.7.2

# 🧩 Optional (agents, experiments, local orchestration)
tqdm==4.66.4
httpx==0.27.0
aiohttp==3.10.10
